Fix downloader to be more efficient
* Fix single httpx client support.

* If file exists abort download to prevent errors with github

*

+++ src/downloader.py
@@ -1,27 +1,35 @@
+import os
+import shutil
+
+import aiofiles
+import aiofiles.os
 import httpx
 from tqdm import tqdm
-import shutil
-import aiofiles
 
 
-async def download_repo(session: httpx.AsyncClient, author: str, repo: str, path: str) -> str:
+def configure_dirs() -> None:
+    os.makedirs('./dist', exist_ok=True)
+
+
+async def download_repo(client: httpx.AsyncClient, author: str, repo: str, path: str) -> str:
     out_path = f'{path}/{author}_{repo}.zip'
+
+    if await aiofiles.os.path.exists(out_path):
+        print(f' {out_path} exists skipping download ')
+        return out_path
+
     async with aiofiles.open(out_path, 'wb') as downloaded_file:
 
-        'https://github.com/brendenlake/omniglot/raw/master/python/images_evaluation.zip'
+        async with client.stream('GET', f'https://github.com/{author}/{repo}/archive/main.zip', follow_redirects=True) as response:
 
-        async with session as client:
+            with tqdm(total=None, unit_scale=True, unit_divisor=1024, unit='B') as progress_bar:
+                downloaded_bytes = response.num_bytes_downloaded
 
-            async with client.stream('GET', f'https://github.com/{author}/{repo}/archive/main.zip', follow_redirects=True) as response:
+                async for chunk in response.aiter_raw():
+                    await downloaded_file.write(chunk)
 
-                with tqdm(total=None, unit_scale=True, unit_divisor=1024, unit='B') as progress_bar:
+                    progress_bar.update(response.num_bytes_downloaded - downloaded_bytes)
                     downloaded_bytes = response.num_bytes_downloaded
-
-                    async for chunk in response.aiter_raw():
-                        await downloaded_file.write(chunk)
-
-                        progress_bar.update(response.num_bytes_downloaded - downloaded_bytes)
-                        downloaded_bytes = response.num_bytes_downloaded
 
     return out_path
 
@@ -29,5 +37,4 @@ async def download_repo(session: httpx.AsyncClient, author: str, repo: str, path
 def unzip_repo(repo_file: str) -> str:
     repo_path = repo_file.removesuffix('.zip')
     shutil.unpack_archive(repo_file, repo_path)
-
     return repo_path
